Add CHERI tests for memory load: byte, half word, word, and double word.
NOTE: byte, half word, and word tests currently pass.  Double word does
not pass; it's not yet clear if this is a bug in the test or in the CHERI
prototype.

Teach the regression test makefile to ignore regression test errors,
allowing further tests to run once one has failed.