Commit a basic test for the 'teqi' trap instruction running in the boot-time exception vector (BEV=1) -- i.e., with exception handlers are ROM addresses. Check various things, including the status register for appropriate EXL, properly installed EPC, the cause register's exception code and branch delay slot flag.  We'll want to flesh out this test with various other combinations of triggers, branch delay slot use, etc.
Right now, most tests fail, as 'teqi' appears not to be triggering an
exception.  This might be due to a bug in CHERI (broken trap instructions),
or a bug in the unit test.  It similarly failed with the 'syscall'
instruction, albeit with an explicit unimplemented instruction warning from
the simulator.