Starting with the syscall instruction unit test, create a breakpoint instruction unit test.
This test fails as we're overloading the breakpoint instruction for
simulator debugging.  As we'll want breakpoints in CHERI software, that's
arguably a bug, although it will require us to find some other bit of the
ISA to squat in for extra features until we have a CHERI coprocessor to
use.