Add CHERI unit tests for the special-purpose 'HI' and 'LO' registers, which take the output of multiply and divide instructions.  This isn't an arithmetic test, rather, a register implementation test: are they initialised properly, can we preserve them properly during context switches, and do multiply and divide operations actually send their results there.
All tests currently fail, as mfhi and mflo wedge the processor if they
are called before any multiply/divide operations have been started.

If that initial test is disabled, then the ddiv test appears to fail due
to an unexpected remainder value in 'HI'.  This might be an arithmetic
bug made by the test author, as oppposed to the prototype, so worth
applying a second set of eyes to.