First cut at a CHERI unit test for the CP0 "compare" register, which implements a cycle-count based timer interrupt.  This is my first attempt to exercise asynchronous exception sources in the CHERI prototype.
This test currently fails; I'm able to configure the compare register,
enable interrupts, and enter a closed loop waiting for it to fire.
When the cycle counter hits the right time, I get the following
Bluespec warning from the simulator, and no entry to the exception
handler:

cyc 2562 - 9000000040000314 : 200000000
     branch delay slot
     Reg  0 <- 0000000000000000

Warning: top.theCP0_causeUpdate3 -- Enqueuing to a full fifo
cyc 2564 - 900000004000030c : 240064800
     Reg  6 <- 0000000000000a01

cyc 2565 - 9000000040000310 : 01000fffe

It seems likely that, once this hardware bug is fixed, we'll then
encounter software bugs in this test that will need fixing, as I
haven't yet been able to exercise it fully.