Add tests for the MIPS sc and scd (store conditional, store conditional double word) to the CHERI unit test suite.  These tests write out data using sc/scd, and then try to read it back using regular loads.  Test results are based on sc/scd returning 1 (success) and on loading back the correct data.
All these tests currently fail.

(sc/scd are defined as succeeding only when pointed at cached memory, and
our default load address is non-cached -- I do some pointer arithmetic to
find an appropriate bit of address space that is cached, but it could be
that this is incorrect.  Or it could be that sc/scd are under-implemented:
a simulator trace suggests that the address arguments are ignored, and
address 0x0 is written to instead?)