Various improvements to the CP0 count + compare register test:
(1) Allow 10 cycles +/- variation in the time from the desired clock tick
    to the exception handler saving the current cycle counter.

(2) Fix a bug: check the right register for ERET checking.

(3) Properly test for the IP7 flag in the CP0 cause register.

(4) Check that the IP7 flag was cleared by clearing the compare register
    in the exception handler.

All tests now pass except for (4).