Add support for C-language unit tests to the CHERI unit test suite.  Add three basic C tests as starting points:
test_ctemplate.c - simple C unit test template
test_casmgp.c - ensure that if we set a register in inline assembly, it
  is visible on return (use $gp since C is instructed not to use that)
test_cretval.c - check that if we return a value from a test function, it
  is visible on return

All tests pass.

This required some rejiggering of the makefile; further rejiggering is
forthcoming to clean up some known nits with it.