Add a CHERI unit test to see whether our uncached and cached direct maps behave as desired.  This test does this by reading via the cache, writing via the uncached mapping, then reading again via the cache: the result should not be visible.
The test currently fails on both CHERI and gxemul.

On CHERI, it likely fails because the cached region is not caching, or
because the uncached region is not bypassing the cache.  As we believe
the UART interface to work, the former may be more likely.

On gxemul, this fails because gxemul does not implement a cache, being an
ISA rather than hardware emulator.  It may be possible to determine this
by checking the CP0 config register to determine the L1 cache size
(perhaps zero) and L2 cache presence (perhaps not present).  The register
is read during the test, but the test suite doesn't yet check it to try
to determine what behaviour is expected.  As such, disable this test for
gxemul for now.