Add basic unit test for CHERI's memory-mapped counter device.  This device will be used in cache testing, but we want to first establish that it behaves properly before relying on it.
This test is tagged with the 'counterdev' Nose attribute, which is
excluded from gxemul runs.

These tests pass.  The counter device reads as little-endian, so we treat
the values as opaque but non-repeating rather than byte swap in either
MIPS assembly or in the unit test, which is fine for our purposes.