Beginnings of support for executing the CHERI unit test suite from multiple relocation points in the address space:
Mark all raw test 'start' symbols as global so that they can be named by the
linker.

In one case, rename a raw test start symbol from 'test' to 'global'; 'test'
is the symbol used in non-raw tests and jumped to by our mini-loader,
whereas 'start' is intended to actually capture the beginning of execution
on the CPU.  This distinction will become more important with support for
executing CHERI unit tests out of cached memory.