Add a CHERI unit test for cscr - storing a capability to memory.
The CHERI architecture document turns out not to describe capability dword
order in memory, so in my test I have assumed that uperms is at offset 0,
otype at offset 8, base at offset 16, and length at offset 24.  However, a
run of the test suite suggests that Jon read the spec as putting the words
in the opposite order.  We should clarify the spec so we know which
interpretation is correct :-).

I have also assumed big endian storage order for the 64-bit fields -- otype,
base, and length.  However, a run of the test suite appears to show that
at least otype and uperms are being written out as little endian.

At least one of the 64-bit words (possibly length?) may not be being
written to memory, as the fourth dword is being read back unmodified from
its assembled (pre-cscr) value.

As such, the four dword tests fail; the underflow and overflow tests for
adjacent dwords in memory succeed.