Modify our register initialisation test so that, rather than checking for initialisation after reset, it tries to initialise registers after reset.
This test previously passed on CHERI despite the fact that registers were
not initialised because the debugging "shadows" of general-purpose
registers, used to generate output for the test framework, were in fact
initialised.  If registers were used without explicit initialisation in
practice, despite appearing to be initialised to the test framework, they
were filled with 0xaaaaaaaaaaaaaaaa rather than the expected
0x000000000000000.

Not initialising the general-purpose register file is actually much more
consistent with stock MIPS, and this test previously failed with gxemul.
In the new world order, it should now pass.

Note: it would be quite useful if we could make the debugging output
reflect internal state -- one way to accomplish this would be to have
the shadow registers be initialised to something other than 0.