Update cgetlen/csetlen test to check that a capability register's length is replaced with, rather than decremented by, the value passed by the general-purpose register argument to the instruction.
This test previously passed, and should now fail, reflecting our changing
view of how the instruction should work.  While the instruction should
indeed always be asked to decrement, it is rarely the case that the software
has the delta to hand: instead, it tends to have the new value (i.e., size
of a data structure to be referenced by the capability), so accept that
instead.