Add two new memcpy implementations:
cmemcpy() takes two capabilities and performs a capability-aware copy between
them, copying 32-byte aligned chunks using capability registers and everything
smaller as a byte-by-byte copy.

smemcpy() is a C memcpy()-compatible implementation that simply tail calls
cmemcpy().  This version should be renamed memcpy() at some point, however this
requires either fixing all of the callers in the initialization code to support
the C calling conventions, or renaming the old memcpy() implementation and
ensuring that everything uses it.

The capability-aware memcpy is tested in the following cases:

- 32-byte aligned start and end
- Aligned start, unaligned end
- Unaligned start, aligned end
- Unaligned start and end
- Different alignment on source and destination

In all but the final case, capabilities that are correctly aligned in the
middle of the copied region are correctly copied, preserving their tag bits.
In the final case, capabilities change alignment and therefore lose their tag
bits.