cheri2: Add associativity to the level 1 caches. This is configurable via the typedefs in CHERITypes (both currently set to one way i.e. direct mapped).
Also update the CP0 config register to reflect this, but note that we
cannot set the line size to its true value of 8 bytes because FreeBSD
does not support this. That doesn't matter because FreeBSD only needs
this information for cache invalidation and cheri2 handles that in
hardware.

Also add a test for access to addresses which alias (at least for a 4k
way size as on cheri2).