Removed "cgetcause" from the floating point exception tests. This made sense for the capability exception tests (where we're interested in the error code set by the capability unit) but not for floating point tests. FPU tests may be run in configurations that don't have a capability unit.
Changed the "comparison with NaN" test to enable the "Invalid" exception.
On closer reading of the MIPS specification, an exception is only raised
on comparison with NaN if the Invalid exception is enabled.

Annex F of the C99 standard says that the relational operators <, > etc.
raise invalid if one of the operands are NaN, and so standards-conforming
compilers ought to be compiling <, > etc. into the trapping MIPS instruction,
at least when #pragma STDC FENV_ACCESS ON is in effect. Neither GCC nor clang
conform to the standard in this respect; clang will warn you that it doesn't
support #pragma STDC FENV_ACCESS ON (and hence isn't going to try
to conform to this part of the standard).