cheritest: Fixes for multicore.
(a) thread_barrier now uses a sync instruction. Not needed on BERI1 or BERI2,
because they have sequentially consistent memory, but in principle needed for
more relaxed memory models we might simulate in future.

(b) At the end of a non-raw test, we check the core/thread id and only halt
the simulation if we're core0. This is so that we don't have a race
condition where it's non-deterministic whether some other core terminates
the simulation before core0 dumps its registers.  A register dump for core0
will be in the trace unless the simulation was terminated for some other
reason, such as reaching the maximum number of clock cycles. Dumps for other
threads/cores may or may not be in the trace (race condition) depending on how
quickly they finish. This check at the end also spefically checks for gxemul,
because we need the tests to work on gxemul for fuzzing (etc.), and gxemul does
not have the BERI-specific CoreId and ThreadId registers.

(c) Also regenerated the fuzz tests, because the previous two changes will
change the size of init.s and lib.s.



Former-commit-id: 989b2815fdc221dd9b3d0546fe62b27815fdf9e2