cheritest: Tagged the tests for csettype with @attr('csettype').
csettype is not in the latest version of the Architecture document, and
is eventually going away. However, it was retained in BERI1/BERI2 so that
old kernels could continue to run, and the argument was made that we should
keep on testing the instruction for as long as the hardware is believed to
support it. Tagging the tests makes it easier to phase out this instruction.

The L3 model does not support csettype.



Former-commit-id: 01e1fdea84f409bef0c4f84f42baa4c0d89b65b3