cheritest: Change L3_NOSEPRED so that the test for the BERI-specific EInstr register is expected to work in the L3 model.
At the moment, this test will fail, because the L3 model doesn't implement this
register. It will eventually be needed, because CheriBSD will probably be
modified to use this register (so that the kernel can work out which
instruction caused an exception, even if the TLB entry for the page containing
the instruction has been replaced), and we want to be able to run CheriBSD in
the L3 model.



Former-commit-id: 02fca2284c7eae6f75a46c15dc6bb7385878a429