cheritest: Changed to be a bit more consistent about the names used for the XML files holding test results.
In particular:

(a) Is it "nosetest" (singular) or "nosetests" (plural)?
(b) Are the components of the file name separated by hyphens or underscores?
(c) Does the name of the system under test (e.g. l3) come before or after "nosetests"?

(The inconsistency is probably the result of different parts of the Makefile having been
written by different people).



Former-commit-id: db798f4be4f19182a6c7267076a411f67b64c0ad