test_raw_fpu_infinity_single: Fixed the test so that floating point 1.0 (rather than integer 1) is added/subtracted/multiplied etc. to infinity.
If the integer value 1 is read as if it were a floating point value, it
is a denormalized floating point number. MIPS R4000 didn't support
denormalized numbers in hardware, so the result is (probably) undefined by
the MIPS spec, even if IEEE 754 defines the result in this case.

The difference didn't matter when BERI was using Altera megafunctions for
floating point multiply/divide, as the megafunctions did the IEEE 754
"right thing". Our new implementation of floating point multiply in Bluespec
is less forgiving about corner cases that MIPS R4000 doesn't require to
be handled.



Former-commit-id: 571ef842986651dc7e63f34840b51c0b919c83a6