Add a test to check whether the delay slot of ccallfast can access IDC.
In hardware, it might be difficult to make ccallfast not have a delay
slot or delay installing the unsealed data cap into IDC by one
instruction. Therefore, the unsealed data cap of the callee is visible
to the caller in IDC in the delay slot. A simple solution is to forbid
the caller to access IDC in the delay slot. Such an access should
fail, which passes the test.




Former-commit-id: 0ef905fc68d37fde17a6f5a18f62171a69149c75