Remove NOT_FLAT=1 from Jenkins build
It speeds up building the simulator but it seems like the simulator is
slower then. I am planning on running the tests twice (binutils and clang)
so this may actually lose us time overall. Also the simulator is pulled in
by other jobs so we should try to have it as fast as possible.