Make the ld/st_unalign tests work with both unaligned enabled and disabled
I added a BERITestBaseClasses.UnalignedLoadStoreTestCase() base class
that ensures that the load/store unaligned tests work both for emulators
with unaligned load emulation and without.

The test case was also relying on the bytes value being aligned to a
cacheline but not being explicit. Fixed by adding a .p2align