Add a DEFINE_COUNTING_CHERI_TRAP_HANDLER macro
This can be used in tests that need to check exceptions instead of
copy-pasting the same handler over and over again. The handler will
count the number of exceptions and save the cause/capcause values.
It will also exit the test function when it receives a syscall trap
to allow running tests in usermode easily.

I also added `set_mips_bev0_handler handler` to simplify setting up the
handler.