Add PYTHON_TESTRUNNER=pytest option to use pytest instead of nosetests
Nosetests (and python2.7) are mostly deprecated and end-of-life so we
should migrate to using pytest ASAP. Fortunately pytest is mostly a drop-in
replacement for nosetests (as long as you have pytest-attrib installed) so
all that needs changing is that the XUnit output flag is different and the
attrib selector is "-a" instead of "-A".

Once all jenkins jobs have been update to use pytest I will make this the
default behaviour.