Check whether a speculative CCallFast can have architectural effect.
There is a bug in the hardware implementation that, if a CCallFast
instruction is the last one in a mispredicted branch sequence, it could
set the ``IDC delay slot" flag for the next instruction at the real
branch target, which means the real branch target with a legal access to
IDC could trigger weird exceptions.

Many thanks to Lawrence who reported this kind of spurious exceptions.