Extract most of test_cp2_x_jr_imprecise test into a template
This test doesn't actually do what the name suggests: it checks that a jr
going out of bounds of $pcc will set epcc to point to the jr instruction
rather than the target.
This should also be tested for jalr,jal,b,bne, etc to ensure that all cases
in QEMU are handled (this is needed because QEMU implementation of branches
is not designed to work with CHERI).