Bump minimum pytest version to 3.6.0
Turns out that before 3.6.0 using pytest.skip() on a class would also apply
it to all methods in the superclass. This caused the sail tests to fail in
jenkins but not on Ubuntu 16.04 without updated pytest because @attr(cached)
was being applied to BeriBaseTestCase.test_exception_count(). This caused
the test that checks no traps were caused to be skipped for all classes.

According to the pytest changelog this was fixed by pull request 3317 which
is included in 3.6.0