Improve cllc_span test
QEMU detects a csc between cllc/cscc but normal byte stores are not detected.
I don't like adding new @attrs but for now just add llscspan_csc for the
test that does a csc before the cscc. Furthermore extend the test to also
do a `sd` between the cllc/cscc which is not detected by QEMU.