cloadtags: better conformance to specification
Attempt to be agnostic to the cacheline width by probing for the reach
of cloadtags as suggested in the ISA doc.  Then, armed with this probe
value, compute the length necessary for a cloadtags and check that a
fetch through the restricted version works.  While here, common up
mostly duplicated code in favor of another assertion.