Add a sim_boot_stub.elf mini bootloader to allow running the ELFs directly
We can no longer map the binaries at the boot address (0xff40000000) since
that is not accessible in QEMU and instead must be a physical address that
happens to overlap with DRAM.

The solution to this is having a tiny bootloader (similar to miniboot) at
0xff40000000. The bootloader fetches the entry point address from an ELF
file and then transfers control there. This more closely matches what the
hardware does: we now always copy the ELF file to address 0x100000.